pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmforge"
version = "0.1.0"
description = "Terminal-based local LLM developer toolkit with live profiling, A/B comparison, and experiment tracking"
readme = "README.md"
requires-python = ">=3.11"
license = "PolyForm-Noncommercial-1.0.0"
authors = [{ name = "LLM Forge Contributors" }]

dependencies = [
    "textual>=1.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "psutil>=6.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tomli>=2.0.0; python_version < '3.12'",
    "tomli-w>=1.0.0",
    "pydantic>=2.0.0",
    "watchfiles>=1.0.0",
    "huggingface-hub>=0.25.0",
]

[project.optional-dependencies]
scoring = [
    "nltk>=3.9.0",
    "rouge-score>=0.1.2",
]
llamacpp = [
    "llama-cpp-python>=0.3.0",
]
rag = [
    "pymupdf>=1.24.0",
]
server = [
    "uvicorn>=0.30.0",
    "starlette>=0.40.0",
    "sse-starlette>=2.0.0",
]
all = ["llmforge[scoring,llamacpp,rag,server]"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "textual-dev>=1.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
llmforge = "llmforge.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
